refactor: news validation
