+++ src/server/news/create.ts
@@ -1,11 +1,11 @@
 import 'server-only'
 import { defaultNewsArticleOldCutoff, newsArticleRealtionsIncluder } from './ConfigVars'
-import { createNewsArticleValidation } from './schema'
+import { createNewsArticleValidation } from './validation'
 import { prismaCall } from '@/server/prismaCall'
 import prisma from '@/prisma'
 import { readCurrenOmegaOrder } from '@/server/omegaOrder/read'
 import { createArticle } from '@/server/cms/articles/create'
-import type { CreateNewsArticleType } from './schema'
+import type { CreateNewsArticleTypes } from './validation'
 import type { ExpandedNewsArticle } from './Types'
 
 /**
@@ -17,7 +17,7 @@ import type { ExpandedNewsArticle } from './Types'
  * @param config.name - The name of the news article (and article)
  * @returns
  */
-export async function createNews(rawdata: CreateNewsArticleType): Promise<ExpandedNewsArticle> {
+export async function createNews(rawdata: CreateNewsArticleTypes['Detailed']): Promise<ExpandedNewsArticle> {
     const { name, description, endDateTime } = createNewsArticleValidation.detailedValidate(rawdata)
 
     const backupEndDateTime = new Date()
