+++ src/server/news/update.ts
@@ -1,13 +1,13 @@
 import 'server-only'
-import { updateNewsArticleValidation } from './schema'
+import { updateNewsArticleValidation } from './validation'
 import { prismaCall } from '@/server/prismaCall'
 import prisma from '@/prisma'
-import type { UpdateNewsArticleType } from './schema'
+import type { UpdateNewsArticleTypes } from './validation'
 import type { SimpleNewsArticle } from '@/server/news/Types'
 
 export async function updateNews(
     id: number,
-    rawdata: UpdateNewsArticleType
+    rawdata: UpdateNewsArticleTypes['Detailed']
 ): Promise<Omit<SimpleNewsArticle, 'coverImage'>> {
     const data = updateNewsArticleValidation.detailedValidate(rawdata)
     return await prismaCall(() => prisma.newsArticle.update({
